requirements.txt
jsonobject-couchdbkit>=0.6.5.2
jsonobject>=0.2.10
sqlalchemy
alembic
psycopg2
celery>=3.0.0
